se verifico la funcionalidad del descuento cuando sea Combo ó Servicio y descuente al stock la cantidad necesaria cuando sea venta ó la aumente cuando sea devolución

+++ PuntoDeVentaV2/Ventana_cancelar_venta.cs
@@ -78,7 +78,7 @@ namespace PuntoDeVentaV2
                                     {
                                         DataRow r_prod_venta = d_prod_venta.Rows[0];
                                         int id_prod = Convert.ToInt32(r_prod_venta["IDProducto"]);
-                                        int cantidad_combo = Convert.ToInt32(r_prod_venta["Cantidad"]);
+                                        decimal cantidad_combo = Convert.ToDecimal(r_prod_venta["Cantidad"]);
 
                                         // Busca los productos relacionados al combo y trae la cantidad para aumentar el stock
                                         DataTable dtprod_relacionados = cn.CargarDatos(cs.productos_relacionados(id_prod));
@@ -87,8 +87,8 @@ namespace PuntoDeVentaV2
                                         {
                                             foreach (DataRow drprod_relacionados in dtprod_relacionados.Rows)
                                             {
-                                                int cantidad_prod_rel = Convert.ToInt32(drprod_relacionados["Cantidad"]);
-                                                int cantidad_prod_rel_canc = cantidad_combo * cantidad_prod_rel;
+                                                decimal cantidad_prod_rel = Convert.ToDecimal(drprod_relacionados["Cantidad"]);
+                                                decimal cantidad_prod_rel_canc = cantidad_combo * cantidad_prod_rel;
 
                                                 cn.EjecutarConsulta($"UPDATE Productos SET Stock = Stock + {cantidad_prod_rel_canc} WHERE ID = {drprod_relacionados["IDProducto"]} AND IDUsuario = {FormPrincipal.userID}");
                                             }
@@ -156,9 +156,9 @@ namespace PuntoDeVentaV2
                                             var concepto = $"DEVOLUCION DINERO VENTA CANCELADA ID {idVenta}";
 
                                             string[] datos = new string[] {
-                                            "retiro", total, "0", concepto, fechaOperacion, FormPrincipal.userID.ToString(),
-                                            efectivo, tarjeta, vales, cheque, transferencia, credito, anticipo
-                                        };
+                                                "retiro", total, "0", concepto, fechaOperacion, FormPrincipal.userID.ToString(),
+                                                efectivo, tarjeta, vales, cheque, transferencia, credito, anticipo
+                                            };
 
                                             cn.EjecutarConsulta(cs.OperacionCaja(datos));
                                         }
